move collab extension
