+++ package.json
@@ -91,8 +91,6 @@
     "zlib": "^1.0.5"
   },
   "dependencies": {
-    "lodash-es": "^4.17.11",
-    "prosemirror-collab": "^1.1.1",
     "socket.io-client": "^2.2.0"
   }
 }
