+++ packages/tiptap-extensions/package.json
@@ -21,10 +21,13 @@
     "url": "https://github.com/scrumpy/tiptap/issues"
   },
   "dependencies": {
+    "lodash-es": "^4.17.11",
     "lowlight": "^1.11.0",
+    "prosemirror-collab": "^1.1.1",
     "prosemirror-history": "^1.0.4",
     "prosemirror-state": "^1.2.2",
     "prosemirror-tables": "^0.7.11",
+    "prosemirror-transform": "^1.1.3",
     "prosemirror-utils": "^0.7.6",
     "prosemirror-view": "^1.8.9",
     "tiptap": "^1.17.0",
